Revert "Compression: add microbenchmark (#9293)"
This reverts commit c3be1a9e0f9883a4a711845863afc621499551d5.
