+++ dbms/src/IO/Compression/tests/gtest_codec_compression.cpp
@@ -1,4 +1,4 @@
-// Copyright 2024 PingCAP, Inc.
+// Copyright 2023 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,9 +14,15 @@
 
 #include <Common/PODArray.h>
 #include <Common/config.h>
-#include <IO/Compression/tests/CodecTestSequence.h>
+#include <DataTypes/DataTypesNumber.h>
+#include <DataTypes/IDataType.h>
+#include <IO/Compression/CompressionFactory.h>
 #include <TestUtils/TiFlashTestBasic.h>
+#include <common/types.h>
 #include <gtest/gtest.h>
+
+#include <magic_enum.hpp>
+#include <random>
 
 
 namespace DB::tests
@@ -24,6 +30,90 @@ namespace DB::tests
 
 template <class T>
 inline constexpr bool is_pod_v = std::is_trivial_v<std::is_standard_layout<T>>;
+
+template <typename T>
+struct AsHexStringHelper
+{
+    const T & container;
+};
+
+template <typename T>
+std::ostream & operator<<(std::ostream & ostr, const AsHexStringHelper<T> & helper)
+{
+    ostr << std::hex;
+    for (const auto & e : helper.container)
+    {
+        ostr << "\\x" << std::setw(2) << std::setfill('0') << (static_cast<unsigned int>(e) & 0xFF);
+    }
+
+    return ostr;
+}
+
+template <typename T>
+AsHexStringHelper<T> AsHexString(const T & container)
+{
+    static_assert(
+        sizeof(container[0]) == 1 && is_pod_v<std::decay_t<decltype(container[0])>>,
+        "Only works on containers of byte-size PODs.");
+
+    return AsHexStringHelper<T>{container};
+}
+
+template <typename T>
+std::string bin(const T & value, size_t bits = sizeof(T) * 8)
+{
+    static const uint8_t MAX_BITS = sizeof(T) * 8;
+    assert(bits <= MAX_BITS);
+
+    return std::bitset<sizeof(T) * 8>(static_cast<uint64_t>(value)).to_string().substr(MAX_BITS - bits, bits);
+}
+
+template <typename T>
+const char * type_name()
+{
+#define MAKE_TYPE_NAME(TYPE)               \
+    if constexpr (std::is_same_v<TYPE, T>) \
+    return #TYPE
+
+    MAKE_TYPE_NAME(UInt8);
+    MAKE_TYPE_NAME(UInt16);
+    MAKE_TYPE_NAME(UInt32);
+    MAKE_TYPE_NAME(UInt64);
+    MAKE_TYPE_NAME(Int8);
+    MAKE_TYPE_NAME(Int16);
+    MAKE_TYPE_NAME(Int32);
+    MAKE_TYPE_NAME(Int64);
+    MAKE_TYPE_NAME(Float32);
+    MAKE_TYPE_NAME(Float64);
+
+#undef MAKE_TYPE_NAME
+
+    return typeid(T).name();
+}
+
+template <typename T>
+DataTypePtr makeDataType()
+{
+#define MAKE_DATA_TYPE(TYPE)               \
+    if constexpr (std::is_same_v<T, TYPE>) \
+    return std::make_shared<DataType##TYPE>()
+
+    MAKE_DATA_TYPE(UInt8);
+    MAKE_DATA_TYPE(UInt16);
+    MAKE_DATA_TYPE(UInt32);
+    MAKE_DATA_TYPE(UInt64);
+    MAKE_DATA_TYPE(Int8);
+    MAKE_DATA_TYPE(Int16);
+    MAKE_DATA_TYPE(Int32);
+    MAKE_DATA_TYPE(Int64);
+    MAKE_DATA_TYPE(Float32);
+    MAKE_DATA_TYPE(Float64);
+
+#undef MAKE_DATA_TYPE
+
+    assert(false && "unknown datatype");
+    return nullptr;
+}
 
 template <typename T, typename Container>
 class BinaryDataAsSequenceOfValuesIterator
@@ -148,6 +238,116 @@ template <typename ContainerLeft, typename ContainerRight>
     }
 }
 
+
+struct CodecTestSequence
+{
+    std::string name;
+    std::vector<char> serialized_data;
+    DataTypePtr data_type;
+    UInt8 type_byte;
+
+    CodecTestSequence(std::string name_, std::vector<char> serialized_data_, DataTypePtr data_type_, UInt8 type_byte_)
+        : name(name_)
+        , serialized_data(serialized_data_)
+        , data_type(data_type_)
+        , type_byte(type_byte_)
+    {}
+
+    CodecTestSequence & append(const CodecTestSequence & other)
+    {
+        assert(data_type->equals(*other.data_type));
+
+        serialized_data.insert(serialized_data.end(), other.serialized_data.begin(), other.serialized_data.end());
+        if (!name.empty())
+            name += " + ";
+        name += other.name;
+
+        return *this;
+    }
+};
+
+CodecTestSequence operator+(CodecTestSequence && left, const CodecTestSequence & right)
+{
+    return left.append(right);
+}
+
+template <typename T>
+CodecTestSequence operator*(CodecTestSequence && left, T times)
+{
+    std::vector<char> data(std::move(left.serialized_data));
+    const size_t initial_size = data.size();
+    const size_t final_size = initial_size * times;
+
+    data.reserve(final_size);
+
+    for (T i = 0; i < times; ++i)
+    {
+        data.insert(data.end(), data.begin(), data.begin() + initial_size);
+    }
+
+    return CodecTestSequence{
+        left.name + " x " + std::to_string(times),
+        std::move(data),
+        std::move(left.data_type),
+        sizeof(T)};
+}
+
+std::ostream & operator<<(std::ostream & ostr, const CompressionMethodByte method_byte)
+{
+    ostr << "Codec{name: " << magic_enum::enum_name(method_byte) << "}";
+    return ostr;
+}
+
+std::ostream & operator<<(std::ostream & ostr, const CodecTestSequence & seq)
+{
+    return ostr << "CodecTestSequence{"
+                << "name: " << seq.name << ", type name: " << seq.data_type->getName()
+                << ", data size: " << seq.serialized_data.size() << " bytes"
+                << "}";
+}
+
+template <typename T, typename... Args>
+CodecTestSequence makeSeq(Args &&... args)
+{
+    std::initializer_list<T> vals{static_cast<T>(args)...};
+    std::vector<char> data(sizeof(T) * std::size(vals));
+
+    char * write_pos = data.data();
+    for (const auto & v : vals)
+    {
+        unalignedStore<T>(write_pos, v);
+        write_pos += sizeof(v);
+    }
+
+    return CodecTestSequence{
+        (fmt::format("{} values of {}", std::size(vals), type_name<T>())),
+        std::move(data),
+        makeDataType<T>(),
+        sizeof(T)};
+}
+
+template <typename T, typename Generator, typename B = int, typename E = int>
+CodecTestSequence generateSeq(Generator gen, const char * gen_name, B Begin = 0, E End = 10000)
+{
+    const auto direction = std::signbit(End - Begin) ? -1 : 1;
+    std::vector<char> data(sizeof(T) * (End - Begin));
+    char * write_pos = data.data();
+
+    for (auto i = Begin; std::less<>{}(i, End); i += direction)
+    {
+        const T v = static_cast<T>(gen(i));
+
+        unalignedStore<T>(write_pos, v);
+        write_pos += sizeof(v);
+    }
+
+    return CodecTestSequence{
+        (fmt::format("{} values of {} from {}", (End - Begin), type_name<T>(), gen_name)),
+        std::move(data),
+        makeDataType<T>(),
+        sizeof(T)};
+}
+
 CompressionCodecPtr makeCodec(const CompressionMethodByte method_byte, UInt8 type_byte)
 {
     CompressionSetting setting(method_byte);
@@ -216,6 +416,99 @@ try
     DB::tests::testTranscoding(*codec, sequence);
 }
 CATCH
+
+///////////////////////////////////////////////////////////////////////////////////////////////////
+// Here we use generators to produce test payload for codecs.
+// Generator is a callable that can produce infinite number of values,
+// output value MUST be of the same type as input value.
+///////////////////////////////////////////////////////////////////////////////////////////////////
+
+auto SameValueGenerator = [](auto value) {
+    return [=](auto i) {
+        return static_cast<decltype(i)>(value);
+    };
+};
+
+auto SequentialGenerator = [](auto stride = 1) {
+    return [=](auto i) {
+        using ValueType = decltype(i);
+        return static_cast<ValueType>(stride * i);
+    };
+};
+
+template <typename T>
+using uniform_distribution = typename std::conditional_t<
+    std::is_floating_point_v<T>,
+    std::uniform_real_distribution<T>,
+    typename std::conditional_t<is_integer_v<T>, std::uniform_int_distribution<T>, void>>;
+
+
+template <typename T = Int32>
+struct MonotonicGenerator
+{
+    explicit MonotonicGenerator(T stride_ = 1, T max_step = 10)
+        : prev_value(0)
+        , stride(stride_)
+        , random_engine(0)
+        , distribution(0, max_step)
+    {}
+
+    template <typename U>
+    U operator()(U)
+    {
+        prev_value = prev_value + stride * distribution(random_engine);
+        return static_cast<U>(prev_value);
+    }
+
+private:
+    T prev_value;
+    const T stride;
+    std::default_random_engine random_engine;
+    uniform_distribution<T> distribution;
+};
+
+template <typename T>
+struct RandomGenerator
+{
+    explicit RandomGenerator(
+        T seed = 0,
+        T value_min = std::numeric_limits<T>::min(),
+        T value_max = std::numeric_limits<T>::max())
+        : random_engine(static_cast<uint_fast32_t>(seed))
+        , distribution(value_min, value_max)
+    {}
+
+    template <typename U>
+    U operator()(U)
+    {
+        return static_cast<U>(distribution(random_engine));
+    }
+
+private:
+    std::default_random_engine random_engine;
+    uniform_distribution<T> distribution;
+};
+
+// auto RandomishGenerator = [](auto i) {
+//     using T = decltype(i);
+//     double sin_value = sin(static_cast<double>(i * i)) * i;
+//     if (sin_value < std::numeric_limits<T>::lowest() || sin_value > static_cast<double>(std::numeric_limits<T>::max()))
+//         return T{};
+//     return static_cast<T>(sin_value);
+// };
+
+auto MinMaxGenerator = []() {
+    return [step = 0](auto i) mutable {
+        if (step++ % 2 == 0)
+        {
+            return std::numeric_limits<decltype(i)>::min();
+        }
+        else
+        {
+            return std::numeric_limits<decltype(i)>::max();
+        }
+    };
+};
 
 // Makes many sequences with generator, first sequence length is 0, second is 1..., third is 2 up to `sequences_count`.
 template <typename T, typename Generator>
@@ -389,18 +682,6 @@ INSTANTIATE_TEST_CASE_P(
             generateSeq<UInt16>(G(RandomGenerator<UInt16>(0))),
             generateSeq<UInt32>(G(RandomGenerator<UInt32>(0, 0, 1000'000'000))),
             generateSeq<UInt64>(G(RandomGenerator<UInt64>(0, 0, 1000'000'000))))));
-
-
-INSTANTIATE_TEST_CASE_P(
-    RepeatInt,
-    CodecTest,
-    ::testing::Combine(
-        IntegerCodecsToTest,
-        ::testing::Values(
-            generateSeq<UInt8>(G(RepeatGenerator<UInt8>(0))),
-            generateSeq<UInt16>(G(RepeatGenerator<UInt16>(0))),
-            generateSeq<UInt32>(G(RepeatGenerator<UInt32>(0))),
-            generateSeq<UInt64>(G(RepeatGenerator<UInt64>(0))))));
 
 // INSTANTIATE_TEST_CASE_P(
 //     RandomishInt,
